Fixed bugs with is_reply function
Added a rule that if the tweet's conversation_id is not the same as the tweet's id_str, it is a reply.

Fixed a bug with the rule that checks to see if  users replied to are in the tweet text.  Previously, it was comparing case sensitively, so it didn't match all cases.